requirements.txt
Flask==1.1.2
h5py==2.10.0
Keras==2.4.3
numpy==1.19.4
gunicorn==20.0.4
Pillow==8.0.1
scipy==1.5.4
tensorflow-cpu==2.4.0